Support for async producer
The Java/Scala Kafka client supports a mechanism for sending
messages asynchronously by using a queue and a thread.
Messages are put on the queue and the worker thread keeps sending
it to the broker.

This ticket implements this feature in python
We use multiprocessing instead of threads to send the messages