Add retry and queue size on async producer
Add retry on async producer in case of send failure. Default set to 3
retry with a back retry set to 300 ms

Add capabilities to limit the size of the async queue in order to avoid
having too much memory consumption by this queue if the async process
is not able to send messages fastly.