Return future from commit_offsets_async
The callers of commit_offsets_async seem to imply it should return a future
(eg. future = commit_offsets_async(...)), but it does not. This causes a mismatch
between KafkaConsumer.commit_async()'s documentation and its behaviour.

This fixes that by returning futures that were already available,
we just had to actually return them.