Bump failing test iterations to 100--is travis happy?
Locally I can Kafka 1.0.1 with python 3.7 to pass intermittently.

On Travis it's not passing. So wondering if I just need more iterations.

Hopefully we only need 20-50, but trying 90 here to see if that truly
fixes it without actually going over the 100 messages we sent... ie, cn
we fetch from both partitions or does it force us to do 100 messages
from first partition before going onto second partition (which would be
a problem)??

Note: not guaranteed that 90 cycles isn't getting us all the 100
messages in the first partition since `fetch_max_bytes=300`... but a
shot in the dark here... if Travis still fails even with this, then may
be something else is the problem.