Merge two tests that are very similar
Previously the `test_kafka_consumer_max_bytes_simple()` was seeing
occasional test failures because it was doing only 10 iterations. And
much of the purpose of it was gutted when Kafka 0.11 came out and
changed the behavior. So this merges the two tests into one which should
be relatively straightforward.

Further discussion in https://github.com/dpkp/kafka-python/pull/1886/files#r316860737