Use OffsetFetchRequest_v3 in _send_offset_fetch_request
Due to a bug in kafka, empty response can return instead of an error for OffsetFetchRequest_v1, since v1 does not support top-level errors.
This issue resulted in an incident where the consumer interpret the bad response as "the consumer group does not have offsets for the topic partition", and therefore it reset the offsets according to the auto-reset-policy which was set to earliest.

In order to avoid the problem, this fix makes sure to use v3, and to check for top-level errors in the response.